requirements.txt
Babel==2.7.0
dj-database-url==0.5.0
Django==2.2.5
django-phonenumber-field==3.0.1
gunicorn==19.9.0
phonenumbers==8.10.19
psycopg2==2.8.4
pytz==2019.2
sqlparse==0.3.0
